Ryddet litt i vedtaksservice (fjernet en del kode vi ikke bruker) (Fortsatt tester som feiler)
