+++ src/main/kotlin/no/nav/familie/ef/sak/service/steg/JournalførVedtaksbrevSteg.kt
@@ -2,7 +2,7 @@ package no.nav.familie.ef.sak.service.steg
 
 import no.nav.familie.ef.sak.repository.domain.Behandling
 import no.nav.familie.ef.sak.service.BehandlingService
-import no.nav.familie.ef.sak.service.VedtaksbrevService
+import no.nav.familie.ef.sak.service.VedtaksbrevIverksettingService
 import no.nav.familie.ef.sak.task.DistribuerVedtaksbrevTask
 import no.nav.familie.kontrakter.felles.journalpost.Journalposttype
 import no.nav.familie.prosessering.domene.TaskRepository
@@ -13,13 +13,13 @@ import org.springframework.stereotype.Service
 
 @Service
 class JournalførVedtaksbrevSteg(private val taskRepository: TaskRepository,
-                                private val vedtaksbrevService: VedtaksbrevService,
+                                private val vedtaksbrevIverksettingService: VedtaksbrevIverksettingService,
                                 private val behandlingService: BehandlingService) : BehandlingSteg<Void?> {
 
     private val logger: Logger = LoggerFactory.getLogger(this::class.java)
 
     override fun utførSteg(behandling: Behandling, data: Void?) {
-        val journalpostId = vedtaksbrevService.journalførVedtaksbrev(behandling.id)
+        val journalpostId = vedtaksbrevIverksettingService.journalførVedtaksbrev(behandling.id)
         behandlingService.leggTilBehandlingsjournalpost(journalpostId
                                                         ?: error("Feil ved journalføring av vedtaksbrev"),
                                                         Journalposttype.U,
