+++ src/test/kotlin/no/nav/familie/ef/sak/repository/VedtaksbrevRepositoryTest.kt
@@ -23,7 +23,8 @@ internal class VedtaksbrevRepositoryTest : OppslagSpringRunnerTest() {
                                       saksbehandlerBrevrequest = "testhallo",
                                       brevmal = "brevmalnavn",
                                       beslutterPdf = null,
-                                      besluttersignatur = null)
+                                      saksbehandlersignatur = "Sakliga Behandlersen",
+                                      besluttersignatur = "")
 
         vedtaksbrevRepository.insert(vedtaksbrev)
